Add a hiera-grep command.
This uses the grepcmd option to search the hiera data repository.

At present we only search the `data/` directory at the root of the
repository (on the assumption that we're working in a control repo
style repository), but the plan is to add support for finding the
correct search paths from the repository hiera config (if any), and
the repository environment config module path (if any).