Update functions and resources.
Mostly to capture the stdlib functions, and explicitly split out the
non-core resources. Mostly this doesn't change the functionality, but it
should make this a bit easier to maintain longer term.

Also a minor update to the function highlighter to capture the chained
function call format (so foo.each - this is universal to all functions
in puppet, so this is worth supporting properly).