Tweak and reorder highlighters.
The order in which highlighters are defined matters more than I
realised, as does the order in which matches are defined in a regex. The
/last/ highlighter that matches wins, and the /first/ match in the regex
wins - thus putting `<|` after `<` in the list of operator matches will
result in only the `<` being matched, with the trailing `|` not matching
and thus not getting the correct face applied. Likewise, if the last
highlighter to match a capitalised type name happens to be the
resourceref highlighter, that's what will set the face.

The gist of all this is that I've now adjusted the order of things to
make sure that the most specific highlighter or match wins, thereby
giving a more consistent colour scheme. It's already shown an erroneous
use of a reserved word in the code I was testing with . . .